+++ lib/recourse.rb
@@ -4,6 +4,7 @@ require 'ransack'
 require 'unicon'
 
 require_relative 'recourse/version'
+require_relative 'recourse/aggregate'
 require_relative 'recourse/bookmarks'
 require_relative 'recourse/colors'
 require_relative 'recourse/themes'